# lomix-next/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lomix Mobile API Documentation",
    "version": "1.0.0",
    "description": "API documentation for Lomix Mobile Application"
  },
  "servers": [
    {
      "url": "https://lomix-be-i2rg.vercel.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local server (Development)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/mobile/gift/send-room": {
      "post": {
        "summary": "Odaya hediye gönder",
        "tags": [
          "Mobile Gifts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "giftId",
                  "roomId"
                ],
                "properties": {
                  "giftId": {
                    "type": "integer"
                  },
                  "roomId": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "integer",
                    "default": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hediye başarıyla gönderildi"
          }
        }
      }
    },
    "/api/mobile/gift/send-user": {
      "post": {
        "summary": "Kullanıcıya hediye gönder (özel mesaj geçmişine kaydeder)",
        "tags": [
          "Mobile Gifts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "giftId",
                  "toUserId"
                ],
                "properties": {
                  "giftId": {
                    "type": "integer"
                  },
                  "toUserId": {
                    "type": "integer"
                  },
                  "amount": {
                    "type": "integer",
                    "default": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hediye başarıyla gönderildi"
          }
        }
      }
    },
...
    "/api/mobile/room/detay": {
      "post": {
        "summary": "Oda Detayı Getir",
        "description": "Belirli bir odanın detay bilgilerini getirir.",
        "tags": [
          "Mobile Rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "roomId",
                  "ownerId"
                ],
                "properties": {
                  "roomId": {
                    "type": "string",
                    "description": "Oda ID"
                  },
                  "ownerId": {
                    "type": "integer",
                    "description": "Oda sahibinin kullanıcı ID'si"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Oda detayı başarıyla getirildi"
          }
        }
      }
    },
...
  },
...
}